feat(content-preview): show versions header when previewing old version
