+++ src/elements/content-explorer/README.md
@@ -28,7 +28,7 @@ var ContentExplorer = require('./ContentExplorer').default;
 ```
 
 ### Screenshot
-<img src="https://user-images.githubusercontent.com/1075325/27887154-092a232a-6194-11e7-82f4-697331ac5cbe.png" style="border: 1px solid #e8e8e8" style="border: 1px solid #e8e8e8" width="600" />
+<img src="https://user-images.githubusercontent.com/1075325/27887154-092a232a-6194-11e7-82f4-697331ac5cbe.png" style="border: 1px solid #e8e8e8;" style="border: 1px solid #e8e8e8;" width="600" />
 ---
 
 ### Props
