+++ src/elements/content-preview/preview-header/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { default } from './PreviewHeaderComponent';
